move isElementInViewport to a utils file
