+++ src/components/Tooltip/utils.ts
@@ -1,0 +1,14 @@
+
+export const isElementInViewport = (
+  el: HTMLElement,
+  win = window,
+  docEl = document.documentElement
+): boolean => {
+  const rect = el.getBoundingClientRect()
+  return (
+    rect.top >= 0 &&
+    rect.left >= 0 &&
+    rect.bottom <= (win.innerHeight || docEl.clientHeight) &&
+    rect.right <= (win.innerWidth || docEl.clientWidth)
+  )
+}
